Version 1.0.30.02
Fixed "WinGet List", which was partially broken by v1.0.30. [thanks twwilliams]

Changed "WinGet List" to retrieve all windows on the system rather than the "last found window" when all four window parameters are blank.

Improved "lines most recently executed" so that when a waiting thread is resumed after being interrupted, the waiting command (e.g. WinWait) will be shown at the bottom of the list, which helps reveal "hung" threads. In addition, "lines most recently executed" now shows 400 lines rather than 200.

+++ Source/hotkey.h
@@ -308,8 +308,8 @@ public:
 	static int FindHotkeyContainingModLR(modLR_type aModifiersLR);  //, int hotkey_id_to_omit);
 	//static int FindHotkeyBySC(sc2_type aSC2, mod_type aModifiers, modLR_type aModifiersLR);
 
-	static char *ListHotkeys(char *aBuf, size_t aBufSize);
-	char *ToText(char *aBuf, size_t aBufSize, bool aAppendNewline);
+	static char *ListHotkeys(char *aBuf, int aBufSize);
+	char *ToText(char *aBuf, int aBufSize, bool aAppendNewline);
 };
